fix(appruntime): Add Cleanup() function to Node interface
This commit introduces a `Cleanup()` function to the `Node` interface in the `pkg/appruntime/base` package. This function allows nodes to perform cleanup operations when they are finished executing.

Currently the main use is to release the connection to pg, by manually releasing the connection it solves a lot of shenanigans like chat timeout with no content, gpts list 504, etc. (gpts hits are also recorded in pg)

By the way, removing a lot of redundant `base.SetAppNamespace` calls.

Signed-off-by: Abirdcfly <fp544037857@gmail.com>